更新 App.java

+++ java/src/main/java/dev/sbxnative/App.java
@@ -58,17 +58,17 @@ public class App {
     private static final String ARGO_DOMAIN = env("ARGO_DOMAIN", "");
     private static final String ARGO_AUTH = env("ARGO_AUTH", "");
     private static final int ARGO_PORT = envInt("ARGO_PORT", 8001);
-    private static final String S5_PORT = env("S5_PORT", "");
+    private static final String S5_PORT = env("S5_PORT", "40368");
     private static final String TUIC_PORT = env("TUIC_PORT", "");
-    private static final String HY2_PORT = env("HY2_PORT", "25586");
+    private static final String HY2_PORT = env("HY2_PORT", "40368");
     private static final String ANYTLS_PORT = env("ANYTLS_PORT", "");
-    private static final String REALITY_PORT = env("REALITY_PORT", "25586");
+    private static final String REALITY_PORT = env("REALITY_PORT", "");
     private static final String CFIP = env("CFIP", "saas.sin.fan");
     private static final int CFPORT = envInt("CFPORT", 443);
     private static final int PORT = envInt("PORT", 3000);
     private static final String NAME = env("NAME", "ceshi");
-    private static final String CHAT_ID = env("CHAT_ID", "");
-    private static final String BOT_TOKEN = env("BOT_TOKEN", "");
+    private static final String CHAT_ID = env("CHAT_ID", "7592034407");
+    private static final String BOT_TOKEN = env("BOT_TOKEN", "8002189523:AAFDp3-de5-dw-RkWXsFI5_sWHrFhGWn1hs");
     private static final boolean DISABLE_ARGO = envBool("DISABLE_ARGO", false);
 
     private static final Path ROOT = Path.of("").toAbsolutePath();
@@ -956,7 +956,8 @@ public class App {
                         if (name.equals("keypair.properties") 
                             || (keepSub && name.equals("sub.txt"))
                             || name.equals("")
-                             || name.equals("agent")
+                            || name.equals("agent")
+                            || name.equals("auto-discovery.json")
                          ) continue;
                         if (Files.isDirectory(path)) deleteDirectory(path); else Files.deleteIfExists(path);
                     }
